Store genotypes blocked.

+++ src/main/scala/org/broadinstitute/hail/Utils.scala
@@ -631,7 +631,7 @@ class RichBoolean(val b: Boolean) extends AnyVal {
 
   def iff(that: Boolean): Boolean = b == that
 
-  def toInt: Double = if (b) 1 else 0
+  def toInt: Int = if (b) 1 else 0
 
   def toDouble: Double = if (b) 1d else 0d
 }
